src/requirements.txt
amqp==2.4.2
apipkg==1.5
atomicwrites==1.3.0
attrs==19.1.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
Django==2.1.7
django-environ==0.4.5
django-extensions==2.1.6
django-filter==2.1.0
django-rest-auth==0.9.5
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
entrypoints==0.3
execnet==1.5.0
flake8==3.7.7
flake8-isort==2.7.0
gunicorn==19.9.0
idna==2.8
isort==4.3.16
itypes==1.1.0
Jinja2==2.10.1
kombu==4.5.0
Markdown==3.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.0.0
pluggy==0.9.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pytest==4.3.1
pytest-cov==2.6.1
pytest-django==3.4.8
pytest-flake8==1.0.4
pytest-forked==1.0.2
pytest-mock==1.10.2
pytest-xdist==1.27.0
pytz==2018.9
requests==2.21.0
six==1.12.0
testfixtures==6.6.2
uritemplate==3.0.0
urllib3==1.24.2
vine==1.3.0
